requirements.txt
cmake
easydict~=1.9
numpy==1.19
numba==0.49.0
opencv-python~=4.5.1.48
matplotlib~=3.3.3
scipy~=1.5.4
scikit-image
tqdm
PyYAML
open3d
cython
tensorboardX
shapely==1.8.1
einops
cumm
timm
pygad==2.17.0
gdown
chardet
